% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{AddHoles}
\alias{AddHoles}
\title{Corrupt a clean image.}
\usage{
AddHoles(I, pc)
}
\arguments{
\item{I}{Clean image.}

\item{pc}{Percetage of pixels to corrupt}
}
\value{
A corrupted image \code{corruptedImage} and \code{maskImage} to
record the corrupted locations.
}
\description{
Corrupt a clean image with randomly selected locations. The corrupted
pixcels are set to 0.
}
\examples{
I=lena_crop
I_noise=AddNoise(I,30)
}
