% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{AddNoise}
\alias{AddNoise}
\title{Add noise to a clean image.}
\usage{
AddNoise(I, sigma)
}
\arguments{
\item{I}{Clean image.}

\item{sigma}{Noise level.}
}
\value{
A noisy image of the same size of I.
}
\description{
Add white noise from N(0,sigma^2) to each pixel of clean image I
}
\examples{
sigma=25; I=lena_crop
I_noise=AddNoise(I,sigma)
}
