% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{ImageReCon}
\alias{ImageReCon}
\title{Combine patches of image into full image}
\usage{
ImageReCon(B, mm, nn, m, n, stepsize = 1)
}
\arguments{
\item{B}{An  (mn) by (M-m+1)(N-n+1) matrix.}

\item{mm, nn}{Full image of size (mm by nn).}

\item{m, n}{Size of the small patches (m by n).}

\item{stepsize}{(optional) The stepsize when splicting the image. Default is 1}
}
\value{
A mm by nn matrix
}
\description{
Suppose an image matrix A of size M by N is splited into (M-m+1)(N-n+1)
overlapped small patches of size m by n. Given a matrix B of size (mn) by (M-m+1)(N-n+1)
, where each column of B is a vectorized small patch, this function try to
reconstruct the full image A.
}
