% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{ImageSplit}
\alias{ImageSplit}
\title{Splict image into small patches}
\usage{
ImageSplit(A, m, n, stepsize = 1)
}
\arguments{
\item{A}{An M by N matrix.}

\item{m, n}{Size of the small patches (m by n).}

\item{stepsize}{(optional) The stepsize when splicting the image. Default is 1}
}
\value{
A (mn) by (M-m+1)(N-n+1) matrix, each column of which is a
 vectorized small patch.
}
\description{
Splict an image matrix A of size M by N into overlapped small patches of size m by n.
Total of (M-m+1)(N-n+1) patches.
}
