% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lena}
\alias{lena}
\title{Image data Lena}
\format{A 512 by 512 matrix}
\usage{
lena
}
\description{
Lena is a benchmark testing image for image processing. The 512 by 512 matrix
is created by reading in the png formed image.
}
\keyword{datasets}
