% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{ODCT}
\alias{ODCT}
\title{Overcomplete Discrete Cosine Transform (ODCT) Generating Function.}
\usage{
ODCT(m, p)
}
\arguments{
\item{m}{Number of grid in each base or the size of each atom in the dictionary.}

\item{p}{Number of bases or the size of the dictionary.}
}
\value{
A m by p matrix. Each column is a DCT base.
}
\description{
\code{ODCT} returns a matrix which is used in dictionary learning as an initial dictionary matrix.
}
\examples{
dic=ODCT(64,100)
# generating an initial dictionary of size 100 with each atom/image patch of size 64.


}
